Seperated in bounds check
I'll end up having a long string of is_such_and_such? to check if
a move is valid or not.

Coincidentally, I couldn't get a negative version of the bounds
check built (with &&'s and between 0 and 80 of 24). Strange. Either
the character was permanently locked in place, or could walk out
of bounds. I may fiddle with that later.