Unincluded Curses
The Python in me says to have fewer global namespace functions, so
I opted to access the curses library by way of the curses module
and not by dumping the curses wrapper into the global namespace.

Also, given I may actually have to learn C and the methodology
behind curses in C in order to write this game in Ruby's curses
wrapper library, I'd like to visually keep Curses and Ruby things
distinct.